PKG: Move normalize funcs to utils pkg
