+++ integration_tests/lnn/perceptron/__init__.py
@@ -1,1 +1,1 @@
-from .perceptron_main import init_perceptron, train_dataset, test_perceptron, normalize_input_vectors, print_perceptron, Perceptron
+from .perceptron_main import init_perceptron, train_dataset, test_perceptron, print_perceptron, Perceptron
