+++ integration_tests/lnn/utils/__init__.py
@@ -1,0 +1,1 @@
+from .utils_main import normalize, normalize_input_vectors
